Fix the convert time and modularization the model system

+++ domain/spot/market/entity/agregateTradeList.go
@@ -2,11 +2,11 @@ package entity
 
 // https://developers.binance.com/docs/binance-spot-api-docs/rest-api/market-data-endpoints#compressedaggregate-trades-list
 type MarketAgregateTradeListEntity struct {
-	TradeId      int    
+	TradeId      int64    
 	Price        string
 	Quantity     string
-	FirstTradeId int
-	LastTradeId  int
+	FirstTradeId int64
+	LastTradeId  int64
 	Timestamp    int64
 	IsBuyerMaker bool
 	IsBestMatch  bool
